requirements.txt
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Data Processing
polars>=0.19.0
pyarrow>=14.0.0

# Machine Learning
torch>=2.1.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Time Series
statsmodels>=0.14.0
prophet>=1.1.5

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.1.0

# Database
influxdb-client>=1.38.0
redis>=5.0.0
psycopg2-binary>=2.9.9

# Notifications
slack-sdk>=3.23.0
python-telegram-bot>=20.6

# Scheduling
apscheduler>=3.10.4
celery>=5.3.4

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
loguru>=0.7.2
